+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added support for diagnostics & completions in [Cargo features](https://doc.rust-lang.org/cargo/reference/features.html)
+- Added support for aliased packages using `package` in [Cargo dependencies](https://doc.rust-lang.org/cargo/reference/specifying-dependencies.html#renaming-dependencies-in-cargotoml)
 
 ## `0.4.1` - January 18th, 2025